test: move generic test to BaseControllerCest
